created a new file for runnumg commands
