+++ includes/executor.h
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   executor.h                                         :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: astein <astein@student.42.fr>              +#+  +:+       +#+        */
+/*   By: anshovah <anshovah@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/09/25 17:55:31 by anshovah          #+#    #+#             */
-/*   Updated: 2023/10/28 19:07:46 by astein           ###   ########.fr       */
+/*   Updated: 2023/11/02 13:21:55 by anshovah         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -56,6 +56,11 @@ enum e_pipe_side
 t_bool  execute(t_mbox *mbox);
 void    run_cmd_system(t_mbox *mbox, t_ast *cmd_node);
 
+/* cmd runner */
+void    run_cmd_main(t_mbox *mbox, t_ast *cmd_node);
+t_bool  run_single_builtin(t_mbox *mbox);
+
+
 /* pipes */
 void    setup_use_pipe(t_mbox *mbox, int status);
 void    setup_pipes(t_mbox *mbox, int *cur_pipe);
@@ -68,15 +73,12 @@ t_bool    setup_redir(t_mbox *mbox, t_ast *redir_node);
 int    heredoc(t_mbox *mbox, t_ast *redir_node, int *cmd_in_fd);
 
 /* executor_utils */
-void    load_executor(t_mbox *mbox);
 void	initialize_io(t_mbox *mbox);
 void	print_executor_output(t_mbox *mbox, int i); //TODO: remove at the end
-void    free_executor(t_mbox *mbox);
 int		cmd_counter(t_ast *ast_node);
-// char    *get_cmd_path(t_mbox *mbox, char *cmd, int i);
 void    get_cmd_av(t_mbox *mbox, t_ast *root);
 void	close_process_fds_v2(t_mbox *mbox);
 
-
+char    *get_cmd_path(t_mbox *mbox, char *cmd, int i, t_bool abs);
 void	free_process_v2(t_mbox *mbox);
 #endif
