+++ src/2_executor/utils_builtin.c
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   utils_builtin.c                                    :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: astein <astein@student.42.fr>              +#+  +:+       +#+        */
+/*   By: anshovah <anshovah@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/10/06 18:54:41 by astein            #+#    #+#             */
-/*   Updated: 2023/10/28 22:50:17 by astein           ###   ########.fr       */
+/*   Updated: 2023/11/02 13:19:48 by anshovah         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -60,20 +60,3 @@ t_bool  is_cmd_builtin(t_mbox *mbox, char *cmd)
     return (ft_false);
 }
 
-/**
- * @brief   traverses through the ll and run builtin cmd via corresponding
- *          function pointer
- * 
- * @param   mbox 
- * @param   cmd_node 
- */
-void    run_cmd_builtin(t_mbox *mbox, t_ast *cmd_node)
-{
-    int i;
-    
-    i = -1;
-    while (mbox->executor.builtins[++i].cmd_name)
-        if (str_cmp_strct(mbox->executor.builtins[i].cmd_name,
-            cmd_node->content))
-            mbox->executor.builtins[i].func_name(mbox, cmd_node->right);
-}
